Move CatalogEntityItem to a separate folder (easier to mock)
